refactor(evidence): split evidence type tests
